mobile/requirements.txt
flet==0.22.*
requests~=2.32.3
pillow==11.0.0